// src/betterdiscord/types/discord/webpack.ts
export interface Require {
    <T = any>(id: PropertyKey): T;
    d(target: object, exports: Record<string, () => any>): void;
    c: Record<PropertyKey, Module>;
    m: Record<PropertyKey, RawModule>;
    e(id: PropertyKey): Promise<unknown>;
    l(url: string, onLoad: (event: Event) => void, key: string, id: string): void;
}

export interface Module<T extends any = any> {
    id: PropertyKey,
    exports: T,
    declarations: Record<string, any>;
    loaded: boolean;
}

export type RawModule = ((module: Module, exports: object, require: Require) => void) & {
    // BD specific properties
    __BD__?: {
        runListeners: (module: Module, exports: object, require: Require) => void;
        originalModule: RawModule;
    };
    __early_patched__?: boolean;
    __raw_module__?: () => RawModule;
};

export type ModuleFilter = (exported: any, module: Module, id: PropertyKey) => any;
export type ExportedOnlyFilter = (exported: any) => any;

export type Options = {
    /** Whether the filter should be run on the exports of modules */
    searchExports?: boolean;
    /** Whether the filter should match the default export but return all exports */
    defaultExport?: boolean;
    /** Whether the filter should be run on the default export of modules. Defaults to true. */
    searchDefault?: boolean;
    /** Whether to return a raw module (an object with `declarations`, `exports`, `id`, and `loaded`) */
    raw?: boolean;
    /** Whether to throw an error if the query fails */
    fatal?: boolean;
    /** The id of the first module to search, with the search running normally afterwards if it does not match */
    firstId?: PropertyKey;
    /** A unique id which is used to store the result of the module lookup for faster searches in the future. Will be autogenerated if not specified. */
    cacheId?: string | null;
    /** A filter to run on the found module's declarations, with the query instead returning the first matching declaration */
    declarationFilter?: ExportedOnlyFilter;
};

export type LazyOptions = Options & {
    /** A signal that cancels waiting for the module */
    signal?: AbortSignal;
};

export type ProxyOptions = Omit<Options, "fatal"> & {
    /** Makes the proxy be an object when `true` sometimes needed for react components */
    typeofIsObject?: boolean;
};

export type MangledOptions = Omit<Options, "declarationFilter"> & {
    /** Whether to map declarations instead of exports */
    mapDeclarations?: boolean;
};

export type LazyMangledOptions = Omit<LazyOptions, "declarationFilter"> & {
    /** Whether to map declarations instead of exports */
    mapDeclarations?: boolean;
};

export type BulkQueries = Options & {
    /** The filter to use to find the module */
    filter: ModuleFilter,
    /** Whether an array of all matching modules should be returned */
    all?: boolean,
    /** A mapper to transform the matched module's exports */
    map?: Record<string, ExportedOnlyFilter>;
    /** Whether the mapper should apply to the module's declarations instead of exports */
    mapDeclarations?: boolean;
};

export type WithKeyOptions = Options & {
    /** The module to find the key on */
    target?: any;
};

export type ModuleWithEffect = [
    any[],
    Record<PropertyKey, RawModule>,
    (require: Require) => void
];
export type ModuleWithoutEffect = [
    any[],
    Record<PropertyKey, RawModule>
];

export type DefaultKey = "default" | "A" | "Ay";